Penetration Test Report
{{Engagement_Name}} sought to confirm the security of their digital infrastructure through a penetration test. The objective of this test was to determine if there any known weaknesses or security flaws present within the systems, applications and processes. This report focuses on the {{Project_Name}} networks and systems of {{Engagement_Name}}. 
Hedgehog Security conducted the penetration test between the {{Start_Date}} and {{End_Date}}. The test was carried out by {{Lead_First-Name_Last-Name}}.

Hedgehog Security tested the systems defined within the scope at the requested time and is unable to comment on the security or vulnerabilities that existed prior to or after the testing was performed. All testing is time limited and it might not be possible to fully investigate every issue or find all possible security issues. Hedgehog Security are unable to comment on systems that were outside of the scope of this report or were unavailable at the time of testing or where required access was not provided.